Conflicts:
	IslandFurnitureERPSystem/CustomerWeb/nbproject/project.properties
	IslandFurnitureERPSystem/CustomerWeb/src/java/Member/Gallery.java
	IslandFurnitureERPSystem/CustomerWeb/src/java/Member/PredefinedAttributes.java
	IslandFurnitureERPSystem/CustomerWeb/src/java/Member/SignUp.java
	IslandFurnitureERPSystem/CustomerWeb/web/index.xhtml
	IslandFurnitureERPSystem/CustomerWeb/web/public/Furniture.xhtml
	IslandFurnitureERPSystem/CustomerWeb/web/public/Gallery.xhtml
	IslandFurnitureERPSystem/CustomerWeb/web/public/SignUp.xhtml
	IslandFurnitureERPSystem/CustomerWeb/web/public/set.xhtml
	IslandFurnitureERPSystem/IslandFurnitureERPSystem-ejb/nbproject/jax-ws.xml
	IslandFurnitureERPSystem/IslandFurnitureERPSystem-ejb/src/java/DataSetUp/dataSetUp.java
	IslandFurnitureERPSystem/IslandFurnitureERPSystem-ejb/src/java/SessionBean/OCRM/OCRMSalesForecastModule.java
	IslandFurnitureERPSystem/IslandFurnitureERPSystem-ejb/src/java/SessionBean/OCRM/OCRMSalesForecastModuleLocal.java
	IslandFurnitureERPSystem/IslandFurnitureERPSystem-war/src/java/ManagedBean/OCRM/CreateStoreProductSalesForecastBean.java
	IslandFurnitureERPSystem/IslandFurnitureERPSystem-war/src/java/ManagedBean/OCRM/OCRMSalesForecastBean.java
	IslandFurnitureERPSystem/IslandFurnitureERPSystem-war/src/java/ManagedBean/OCRM/ViewProductSalesForecastBean.java
	IslandFurnitureERPSystem/nbproject/build-impl.xml

+++ IslandFurnitureERPSystem/IslandFurnitureERPSystem-ejb/src/java/Entity/Store/OCRM/TransactionItem.java
@@ -25,7 +25,7 @@ public class TransactionItem implements Serializable {
     private Long TransactionItemId;
     
     private Long itemId;//refer to storeproductId or storeRetailProductId
-    private Double amount;
+    private int amount;
     private Double unitPrice;
     private Double totalPrice;
     private String itemName;
@@ -55,7 +55,7 @@ public class TransactionItem implements Serializable {
         return itemId;
     }
 
-    public Double getAmount() {
+    public int getAmount() {
         return amount;
     }
 
@@ -83,7 +83,7 @@ public class TransactionItem implements Serializable {
         this.itemId = itemId;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(int amount) {
         this.amount = amount;
     }
